pyproject.toml
[tool.poetry]
name = "insurance-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.135.3"
alembic = "^1.18.4"
sqlalchemy = "^2.0.49"
pydantic-settings = "^2.13.1"
yarl = "^1.23.0"
asyncpg = "^0.31.0"
bcrypt = "^5.0.0"
pyjwt = "^2.12.1"
python-multipart = "^0.0.22"
uvicorn = "^0.43.0"
cryptography = "^46.0.6"
ollama = "^0.6.1"
redis = "^7.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
